day6

+++ tensquare_user/src/main/java/com/tensquare/user/controller/AdminController.java
@@ -3,6 +3,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -20,6 +21,8 @@ import entity.Result;
 import entity.StatusCode;
 import util.JwtUtil;
 
+import javax.servlet.http.HttpServletRequest;
+
 /**
  * 控制器层
  * @author Administrator
@@ -36,6 +39,9 @@ public class AdminController {
 	@Autowired
 	private JwtUtil jwtUtil;
 
+	@Autowired
+	private HttpServletRequest request;
+
 	@RequestMapping(value = "/login",method = RequestMethod.POST)
 	public Result login(@RequestBody Map<String,String> loginMap){
 		Admin admin = adminService.findByLoginnameAndPassword(loginMap.get("loginname"),loginMap.get("password"));
@@ -44,7 +50,7 @@ public class AdminController {
 			Map map = new HashMap();
 			map.put("token",token);
 			map.put("name",admin.getLoginname());
-			return new Result(true, StatusCode.OK, "登录成功");
+			return new Result(true, StatusCode.OK, "登录成功",map);
 		}else {
 			return new Result(false,StatusCode.ERROR,"登录失败");
 		}
@@ -121,6 +127,21 @@ public class AdminController {
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
 	public Result delete(@PathVariable String id ){
+		String authHeader = request.getHeader("Authorization");
+		if (authHeader==null){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
+		if (!authHeader.startsWith("Bearer ")){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
+		String token = authHeader.substring(7);
+		Claims claims = jwtUtil.parseJWT(token);
+		if (claims==null){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
+		if (!claims.get("roles").equals("admin")){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
 		adminService.deleteById(id);
 		return new Result(true,StatusCode.OK,"删除成功");
 	}
